Refracted all the code and added comments
