+++ README.md
@@ -9,11 +9,11 @@
 4. If the date string is **valid** the api returns a JSON having the structure 
 `{"unix": <date.getTime()>, "utc" : <date.toUTCString()> }`
 e.g. `{"unix": 1479663089000 ,"utc": "Sun, 20 Nov 2016 17:31:29 GMT"}`.
-5. If the date string is **invalid** the api returns a JSON having the structure `{"unix": null, "utc" : "Invalid Date" }`. It is what you get from the date manipulation functions used above.
+5. If the date string is **invalid** the api returns a JSON having the structure `{"error" : "Invalid Date" }`. It is what you get from the date manipulation functions used above.
 
 #### Example usage:
 * https://curse-arrow.hyperdev.space/api/timestamp/2015-12-15
 * https://curse-arrow.hyperdev.space/api/timestamp/1450137600000
 
 #### Example output:
-* { "unix": 1450137600, "natural": "December 15, 2015" }
+* { "unix": 1450137600, "utc": "Fri, 01 Nov 2019 22:28:39 GMT" }
